requirements.txt
# Auto Study - 自动学习系统依赖包
# Python 3.8+ required

# Web自动化
playwright>=1.40.0

# OCR验证码识别（兼容性版本）
ddddocr>=1.0.6

# 日志管理
loguru>=0.7.0

# 配置管理
python-dotenv>=1.0.0
pyyaml>=6.0

# 终端UI
rich>=13.0.0

# 数据库
sqlalchemy>=2.0.0

# 定时任务
apscheduler>=3.10.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# HTTP请求
httpx>=0.25.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 图像处理
pillow>=10.0.0

# 加密
cryptography>=41.0.0

# 文件监控
watchdog>=3.0.0

# 系统监控
psutil>=5.9.0